requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
instagrapi==1.16.20
Pillow==9.1.1
pycryptodomex==3.14.1
pydantic==1.9.1
PySocks==1.7.1
python-dotenv==0.20.0
pyunsplash==1.0.0rc2
requests==2.28.0
typing_extensions==4.2.0
urllib3==1.26.9
